added getallusers, updateuser and delete user features

+++ src/main/java/com/project/userservice/repository/UserRepository.java
@@ -3,7 +3,6 @@ package com.project.userservice.repository;
 import com.project.userservice.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
 }
-
